requirements.txt
redis>=4.5.1
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
httpx>=0.24.0
asyncio>=3.4.3
aiohttp>=3.8.5
pytest>=7.4.0
pytest-asyncio>=0.21.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 